+++ R/endpoints.R
@@ -253,9 +253,7 @@ verify_result_available <- function(queue, id) {
   task_status <- queue$queue$task_status(id)
   if (task_status == "COMPLETE") {
     result <- queue$result(id)
-    if (!naomi:::is_hintr_output(result)) {
-      hintr_error(t_("UNKNOWN_OUTPUT_TYPE"), "UNKNOWN_OUTPUT_TYPE")
-    }
+    naomi:::assert_model_output_version(result)
   } else if (task_status == "ERROR") {
     result <- queue$result(id)
     trace <- c(sprintf("# %s", id), result$trace)
@@ -290,47 +288,41 @@ plotting_metadata <- function(iso3) {
   )
 }
 
-download_spectrum <- function(queue) {
-  download(queue, "spectrum", "naomi-output", ".zip")
+download_submit <- function(queue) {
+  function(id, type) {
+    verify_result_available(queue, id)
+    ## API path should be - separated but we
+    ## use _ for names in naomi
+    type <- gsub("-", "_", type, fixed = TRUE)
+    tryCatch(
+      list(id = scalar(queue$submit_download(queue$result(id), type))),
+      error = function(e) {
+        hintr_error(e$message, "FAILED_TO_QUEUE")
+      }
+    )
+  }
 }
 
-download_coarse_output <- function(queue) {
-  download(queue, "coarse_output", "coarse-output", ".zip")
-}
-
-download_summary <- function(queue) {
-  download(queue, "summary", "summary-report", ".html")
-}
-
-download <- function(queue, type, filename, ext) {
+download_result <- function(queue) {
   function(id) {
     tryCatch({
       res <- queue$result(id)
-      if (is_error(res)) {
-        hintr_error(res$message, "MODEL_RUN_FAILED")
+      if (is_error(res) || is.null(res$path)) {
+        msg <- res$message
+        if (is.null(msg)) {
+          msg <- t_("FAILED_ADR_METADATA")
+        }
+        hintr_error(msg, "OUTPUT_GENERATION_FAILED")
       }
-      ## We renamed download from "summary" to "coarse_output" to be more
-      ## representative of the actual content of the download and in
-      ## preparation for adding a summary report.
-      ## To be backwards compatible with old model results from the app
-      ## we need to fallback to old name if the new name isn't available in
-      ## the model result.
-      if ("coarse_output_path" %in% names(res)) {
-        coarse_output <- res$coarse_output_path
-      } else {
-        coarse_output <- res$summary_path
-      }
-      path <- switch(type,
-                     "spectrum" = res$spectrum_path,
-                     "coarse_output" = coarse_output,
-                     "summary" = res$summary_report_path)
-      if (is.null(path)) {
-        hintr_error(t_("MODEL_RESULT_OLD",
-                       list(type = gsub("-", " ",
-                                        tools::file_path_sans_ext(filename)))),
-                    "MODEL_RESULT_OUT_OF_DATE")
-      }
-      bytes <- readBin(path, "raw", n = file.size(path))
+      filename <- switch(res$metadata$type,
+                         spectrum = "naomi-output",
+                         coarse_output = "coarse-output",
+                         summary = "summary-report")
+      ext <- switch(res$metadata$type,
+                         spectrum = ".zip",
+                         coarse_output = ".zip",
+                         summary = ".html")
+      bytes <- readBin(res$path, "raw", n = file.size(res$path))
       bytes <- porcelain::porcelain_add_headers(bytes, list(
         "Content-Disposition" = build_content_disp_header(res$metadata$areas,
                                                           filename, ext),