+++ tests/testthat/test-endpoints.R
@@ -14,7 +14,7 @@ test_that("plumber api can be built", {
                   c("options", "submit", "status", "result", "cancel", "debug"))
   expect_setequal(names(api$routes$calibrate),
                   c("options", "submit", "status", "result", "plot"))
-  expect_equal(names(api$routes$meta), "plotting")
+  expect_setequal(names(api$routes$meta), c("plotting", "adr"))
   expect_equal(names(api$routes$`chart-data`), "input-time-series")
 })